Added "How much" question type

See merge request cse1105/2021-2022/team-repositories/oopp-group-76/repository-template!84

+++ client/src/main/java/client/Session.java
@@ -12,7 +12,7 @@ public class Session {
 
     private static String nickname;
     private static String serverAddr;
-    private static int questionNum;
+    private static long questionNum;
 
     public static QuizzQuestionServerParsed emptyQ = new QuizzQuestionServerParsed(new QuizzQuestion("0",new Activity("0","0","0",Long.valueOf(0),"0"),new Activity("0","0","0",Long.valueOf(0),"0"),new Activity("0","0","0",Long.valueOf(0),"0")),-1,-1, new ArrayList<Joker>());
 
@@ -25,7 +25,7 @@ public class Session {
         return Session.serverAddr;
     }
 
-    public static int getQuestionNum() {
+    public static long getQuestionNum() {
         return Session.questionNum;
     }
 
@@ -37,7 +37,7 @@ public class Session {
         Session.serverAddr = serverAddr;
     }
 
-    public static void setQuestionNum(int questionNum) {
+    public static void setQuestionNum(long questionNum) {
         Session.questionNum = questionNum;
     }
 }
